+++ CHANGELOG.md
@@ -8,6 +8,8 @@ e este projeto adere a [Semantic Versioning](https://semver.org/lang/pt-BR/).
 
 ## [1.0.7-beta.0]
 - implementação do PipelineSolucoesLinkTokens.
+- implementação do PipelineSolucoesSafeAreaTokens.
+- implementação do PipelineSolucoesFooterTokens.
 
 ## [1.0.4] - 2025-12-30
 - Ajuste no PipelineSolucoesFieldTokens